Move file and refactor some functionality to new serice
